Code backup for original SpotMicro
Since The writer of this code is drastically redesigning the robot and the code, i made a backup here for all of those who don't switch to the new model.
This code is written by Adham Elarawaby. With some addon's and adjustments from myself for better compatibility with my SpotMICRO Controller board.